% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{colinput}
\alias{colinput}
\title{Column input}
\usage{
colinput(width, id, label, choices, ...)
}
\arguments{
\item{width}{An integer specifying the width of the column.}

\item{id}{A string specifying the ID module.}

\item{label}{A string specifying the label for the select input.}

\item{choices}{A vector of strings specifying the choices of the select input.}

\item{...}{Additional arguments passed to shiny::selectInput().}
}
\value{
A column containing a select input.
}
\description{
This function creates a column containing a select input in a shiny app.
}
