% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\docType{data}
\name{colpalettes}
\alias{colpalettes}
\title{Color palettes}
\format{
An object of class \code{character} of length 15.
}
\usage{
colpalettes
}
\value{
A character vector of color palette names.
}
\description{
This is a character vector containing the names of several color palettes.
}
\keyword{datasets}
