% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{common_input}
\alias{common_input}
\title{Common input}
\usage{
common_input(panel, modns)
}
\arguments{
\item{panel}{A string indicating whether the panel is a "stats" panel or a 
"plotting" panel.}

\item{modns}{A string to namespace inputs to avoid id conflicts when using 
multiple instances of the same module.}
}
\value{
A tag list containing a select input for the treatment and a select 
input for the target variable if the panel is "stats". Otherwise, it returns 
a numeric input for the download image width and a numeric input for the 
download image height if the panel is "plotting".
}
\description{
This function creates the common inputs for each tab. It provides statistical
as well as plotting commands.
}
