% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{compmeans_ui}
\alias{compmeans_ui}
\title{Mean comparison UI}
\usage{
compmeans_ui(id, panel)
}
\arguments{
\item{id}{A character string specifying the namespace for the UI elements.}

\item{panel}{A character string specifying the panel to display in the UI ("stats" or "plotting").}
}
\value{
A Shiny UI object.
}
\description{
This function generates the user interface for the Explore tab of a Shiny application.
}
