% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_explore.R
\name{dist_detect}
\alias{dist_detect}
\title{Detect the distribution of numeric variables in a data.table}
\usage{
dist_detect(d, cutoff = 0.05)
}
\arguments{
\item{d}{A data table}

\item{cutoff}{An optional numeric value specifying the p-value cutoff for determining normality (default is 0.05)}
}
\value{
A data table with the results of the Shapiro-Wilk test for normality for each numeric variable in the input data table
}
\description{
This function takes a data.table and an optional cutoff value as input and returns a data.table with the results of the Shapiro-Wilk test for normality for each numeric variable in the input data.table. The output data.table contains the variable name, W-statistic, p-value, and a column indicating whether the variable is normally distributed or not based on the specified cutoff value.
}
