% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_linear-reg.R
\name{escape_for_latex}
\alias{escape_for_latex}
\title{Escapes special characters for LaTeX output}
\usage{
escape_for_latex(s)
}
\arguments{
\item{s}{a character string to be escaped for LaTeX output}
}
\value{
a character string with all special characters escaped for LaTeX output
}
\description{
This function takes a character string as input and escapes special characters
that are used in LaTeX documents such as: backslash, ampersand, percent, hash,
underscore, curly braces, tilde, and circumflex.
}
