% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plotting_tools.R
\name{gg_corplot}
\alias{gg_corplot}
\title{Generate a correlation plot using gg_corplot}
\usage{
gg_corplot(d)
}
\arguments{
\item{d}{A data frame to generate the plot for.}
}
\value{
A corplot object
}
\description{
This function generates a correlation plot for a given data frame, using the \code{corrplot} package.
}
