% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plotting_tools.R
\name{ggplot_hist}
\alias{ggplot_hist}
\title{Histogram plot}
\usage{
ggplot_hist(d, treatment, colpal = "jco")
}
\arguments{
\item{d}{a data table}

\item{treatment}{a character string representing the name of the treatment variable}

\item{colpal}{an optional character string representing the name of the color palette to use (default is "jco")}
}
\value{
a ggplot object
}
\description{
This function takes a data table, a treatment variable and an optional color palette as inputs
and produces a histogram plot using ggplot2 with overlaid normal density and vertical lines
for the means of each numeric variable, grouped by the treatment variable.
}
