% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plotting_tools.R
\name{ggplot_lm}
\alias{ggplot_lm}
\title{Plotting function for linear models}
\usage{
ggplot_lm(d, model, input)
}
\arguments{
\item{d}{A data.frame}

\item{model}{A linear model object}

\item{input}{A list containing several options for creating the plot. The options are:
\describe{
\item{xvar}{The independent variable}
\item{yvar}{The dependent variable}
\item{colvar}{The grouping variable (optional)}
\item{plot_type}{The type of plot to be created ("model check" or "scatter")}
}}
}
\value{
A ggplot object
}
\description{
This function takes in a data frame, a linear model, and several inputs to create a plot.
The input options are \code{xvar} for the independent variable, \code{yvar} for the dependent variable,
\code{colvar} for the grouping variable (if applicable), and \code{plot_type} for the type of plot
to be created ("model check" or "scatter").
}
\details{
If \code{colvar} is not specified or is set to "none", the grouping variable is set to "colvar"
and all data points are colored the same. Otherwise, the grouping variable is set to the
specified \code{colvar}.

For "model check" plots, the function creates a multiplot of six plots, showing the distribution
of residuals, normal Q-Q plot of residuals, scale-location plot, residuals vs. leverage plot,
cook's distance plot, and residual vs. fitted plot.

For "scatter" plots, the function creates a scatterplot of the data with the regression line overlaid.
If a \code{colvar} is specified, the data points are colored by that variable. Additionally, the
function adds error bars to the plot representing the standard error of the prediction for each data point.
}
