% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plotting_tools.R
\name{ggplot_qq}
\alias{ggplot_qq}
\title{Create a QQ plot using ggplot}
\usage{
ggplot_qq(d, treatment, colpal)
}
\arguments{
\item{d}{A data table containing the data to plot.}

\item{treatment}{A character string specifying the column name of the treatment variable in the data table.
If "none", a dummy treatment variable will be created and used for plotting.}

\item{colpal}{A character string specifying the color palette to use for plotting.
If treatment is "none", this parameter will be set to "black".}
}
\value{
A ggplot object representing a QQ plot of the data.
}
\description{
Create a QQ plot using ggplot
}
\examples{
library(data.table)
d <- data.table(x = rnorm(100), y = rnorm(100), z = rnorm(100), treatment = rep(c("A", "B"), each = 50))
ggplot_qq(d, "treatment", "jco")

}
