% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_linear-reg.R
\name{linreg_module}
\alias{linreg_module}
\title{Linear Regression module for Shiny app}
\usage{
linreg_module(id, d)
}
\arguments{
\item{id}{A string that specifies the module's ID. The ID should be unique
among the app's modules.}

\item{d}{A reactive object containing the data to be analyzed.}
}
\value{
A Shiny module that provides a user interface and server logic for
a linear regression test and related plots.
}
\description{
This module creates a user interface and a server for a linear regression test
and related plots. It also updates its inputs with the corresponding information
from the data file.
}
