% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\docType{data}
\name{mainpanel_txt_ui}
\alias{mainpanel_txt_ui}
\title{Mainpanel Text User Interface}
\format{
An object of class \code{list} of length 4.
}
\usage{
mainpanel_txt_ui
}
\value{
A character vector with the text for each section.
}
\description{
This function contains the text for the four types of analysis tools in the 
shiny app.
}
\keyword{datasets}
