% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plotting_tools.R
\name{mean_sd}
\alias{mean_sd}
\title{Exporting ggpubr::mean_sd function}
\usage{
mean_sd(x, error.limit = "both")
}
\description{
The ggpubr package provides a function called mean_sd, which calculates the mean and standard deviation of a vector. However, in order to use this function within certain ggpubr functions, such as barplot, it is necessary to export it to a function within your own package. This function does not have any parameters or return values, it simply imports the ggpubr::mean_sd function and exports it to your package.
}
