% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_one-way-ANOVA.R
\name{oneway_module}
\alias{oneway_module}
\title{One-Way ANOVA Module}
\usage{
oneway_module(id, d)
}
\arguments{
\item{id}{A character string that specifies the namespace for the module.}

\item{d}{A reactive expression that returns the data to be used in the analysis.}
}
\value{
A module server function that can be called within a Shiny application.
}
\description{
This module performs a one-way analysis of variance (ANOVA) on the given data
and generates a plot of the results.
}
