% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{oneway_ui}
\alias{oneway_ui}
\title{ANOVA One-way UI}
\usage{
oneway_ui(id, panel)
}
\arguments{
\item{id}{A character string specifying the namespace for the UI elements.}

\item{panel}{A character string specifying the panel to display in the UI ("stats" or "plotting").}
}
\value{
A Shiny UI object.
}
\description{
This function generates the user interface for the ANOVA One-way tab of a Shiny application.
}
