% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_post-hoc.R
\name{ph_dunn}
\alias{ph_dunn}
\title{Conducts post-hoc pairwise comparisons after a Kruskal-Wallis test.}
\usage{
ph_dunn(d, treatment, variable, p.adjust.method = "bonferroni")
}
\arguments{
\item{d}{A data frame containing the data to be analyzed.}

\item{treatment}{A string specifying the name of the categorical variable that defines the groups to be compared.}

\item{variable}{A string specifying the name of the continuous variable to be analyzed.}

\item{p.adjust.method}{A string specifying the method used to adjust p-values for multiple comparisons. Default is "bonferroni".}
}
\value{
A data table containing the results of the pairwise comparisons, including the names of the compared groups, the test statistic, the p-value, and any adjustment for multiple testing.
}
\description{
This function conducts post-hoc pairwise comparisons between groups in a 
dataset, after a Kruskal-Wallis test has indicated significant differences 
among the groups. The function uses the Dunn's test to compare all possible 
pairwise differences between the groups.
}
