% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_post-hoc.R
\name{ph_mult_ttest}
\alias{ph_mult_ttest}
\title{Conducts post-hoc pairwise comparisons t-test.}
\usage{
ph_mult_ttest(d, treatment, variable, p.adjust.method = "bonferroni")
}
\arguments{
\item{d}{A data frame containing the data to be analyzed.}

\item{treatment}{A string specifying the name of the categorical variable that defines the groups to be compared.}

\item{variable}{A string specifying the name of the continuous variable to be analyzed.}

\item{p.adjust.method}{A string specifying the method used to adjust p-values for multiple comparisons. Default is "bonferroni".}
}
\value{
A data table containing the results of the pairwise comparisons, including the names of the compared groups, the difference in means, the standard error, the t-value, the degrees of freedom, the p-value, and any adjustment for multiple testing.
}
\description{
This function conducts post-hoc pairwise comparisons between groups in a 
dataset. The function uses the Welch's t-test to compare all possible pairwise 
differences between the groups, and adjusts p-values for multiple comparisons 
using the specified method.
}
