% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plotting_tools.R
\name{plot_lmcoef}
\alias{plot_lmcoef}
\title{Plot coefficients of a linear model}
\usage{
plot_lmcoef(model, input)
}
\arguments{
\item{model}{The linear model object to be plotted.}

\item{input}{A list containing the input values for the function.}
}
\value{
A list containing the ggplot object generated.
}
\description{
This function plots the coefficients of a linear model with their corresponding standard errors and p-values.
}
