% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plotting_tools.R
\name{plot_one_comp_m}
\alias{plot_one_comp_m}
\title{Plot data for one-sample comparison test}
\usage{
plot_one_comp_m(d, input, test_out)
}
\arguments{
\item{d}{A data.frame or data.table object containing the data to plot.}

\item{input}{A list object containing input parameters for the plot. Required 
parameters include: treatment, variable, ref.group, plot_type, and colpal. 
Optional parameters include: paired.}

\item{test_out}{A list object containing the results of a statistical test.}
}
\value{
A ggplot object with the plotted data.
}
\description{
This function generates a plot to compare a single group to a reference value,
using the data and parameters provided.
}
