% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_linear-reg.R
\name{reg_test}
\alias{reg_test}
\title{reg_test}
\usage{
reg_test(d, input)
}
\arguments{
\item{d}{A dataframe containing the data to be analyzed.}

\item{input}{A list containing the following user-specified input parameters:
\describe{
\item{variable}{The dependent variable for the regression analysis.}
\item{treatment}{The independent variable(s) for the regression analysis. Note that this can be more than one.}
\item{cilevel}{The level of confidence for the confidence interval calculation.}
\item{allow_interaction}{A boolean value indicating whether to include interaction terms in the model.}
}}
}
\value{
A list containing the following elements:
\describe{
\item{model}{The regression model object.}
\item{htmlout}{The regression summary in HTML format, as produced by the stargazer package.}
\item{report}{The regression summary in text format, as produced by the reportr package.}
}
}
\description{
A function to perform a regression analysis on a given data.frame, based on user-specified input parameters.
}
