% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_explore.R
\name{summary_table}
\alias{summary_table}
\title{Summarize data by treatment group}
\usage{
summary_table(d, treatment)
}
\arguments{
\item{d}{A data table}

\item{treatment}{A character string specifying the name of the treatment variable in the data table}
}
\value{
A summary table of the numeric variables in the input data table, grouped by the treatment variable
}
\description{
This function takes a data.table and a treatment variable and returns a summary table of the numeric variables in the data.table, grouped by the treatment variable. The summary statistics calculated are: sample size (N), mean, median, variance (Var), and interquartile range (IQR).
}
