% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_two-way-ANOVA.R
\name{twoway_module}
\alias{twoway_module}
\title{Two-way ANOVA Shiny Module}
\usage{
twoway_module(id, d)
}
\arguments{
\item{id}{The module ID.}

\item{d}{The data to use in the analysis.}
}
\description{
This module performs a two-way ANOVA and post-hoc analysis based on user-defined inputs, and
creates a plot to visualize the results.
}
