% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_updater.R
\name{updater_module}
\alias{updater_module}
\title{Update input selectors for statistical analysis modules}
\usage{
updater_module(id, d)
}
\arguments{
\item{id}{The module ID.}

\item{d}{The data to use for updating the input selectors.}
}
\value{
An observer that updates the input selectors when the data or module ID
  changes.
}
\description{
This function updates the input selectors for a statistical analysis module
based on the data provided. The input selectors include options for selecting
the treatment, target variable, reference group, x-variable, y-variable, and
color variable for plotting.
}
