% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plotting_tools.R
\name{add_pvals}
\alias{add_pvals}
\title{Add pvalues}
\usage{
add_pvals(
  ggp,
  test_type = "t.test",
  xvar,
  colvar,
  pval_format = c("p.adj.format", "p.adj.signif"),
  betweet_type = "within",
  p_adjust = "bonferroni",
  test_paired = FALSE,
  dodge = 0.8
)
}
\description{
Add pvalues
}
